test_compression: fix 32-bit and macOS sign-compare errors
The previous fix incorrectly added 0u suffix to two ASSERT_GT sites where
the LHS is ssize_t (signed), which on 32-bit Linux is int. This produced
a new sign-compare error (int vs unsigned int) that CI caught.

- Revert 'ASSERT_GT(compressed_len, 0u)' to 'ASSERT_GT(compressed_len, 0)'
  where compressed_len is ssize_t (lines 187, 297)
- Cast VKCS_ENVELOPE_SIZE to (size_t) where the LHS is size_t (lines 1047,
  1326); the macro is an int literal, so this was int vs unsigned int on
  32-bit.

Verified with a 32-bit Ubuntu docker container (g++-multilib, -m32,
-Werror=sign-compare). test_compression.cpp now compiles cleanly.

+++ src/unit/test_compression.cpp
@@ -184,7 +184,7 @@ TEST_F(CompressionTest, streamCompressDecompressRoundTrip) {
     ssize_t compressed_len = streamCompressFeed(&sc, compressed, bound,
                                                 (const uint8_t *)input, input_len,
                                                 FLUSH_END);
-    ASSERT_GT(compressed_len, 0u) << "compress should succeed";
+    ASSERT_GT(compressed_len, 0) << "compress should succeed";
     ASSERT_EQ(sc.stream_started, false) << "frame should be closed after FLUSH_END";
     streamCompressorDestroy(&sc);
 
@@ -294,7 +294,7 @@ TEST_F(CompressionTest, streamDecompressFeedErrors) {
     ssize_t compressed_len = streamCompressFeed(&sc, compressed, bound,
                                                 (const uint8_t *)payload, strlen(payload),
                                                 FLUSH_END);
-    ASSERT_GT(compressed_len, 0u);
+    ASSERT_GT(compressed_len, 0);
     streamCompressorDestroy(&sc);
 
     ASSERT_TRUE(streamDecompressFeed(&sd, out, sizeof(out),
@@ -1044,7 +1044,7 @@ TEST_F(CompressionTest, compressRioRoundTrip) {
     /* Get the compressed output from the buffer rio */
     sds compressed = buffer_rio.io.buffer.ptr;
     size_t compressed_len = sdslen(compressed);
-    ASSERT_GT(compressed_len, VKCS_ENVELOPE_SIZE) << "compressed output should exist";
+    ASSERT_GT(compressed_len, (size_t)VKCS_ENVELOPE_SIZE) << "compressed output should exist";
 
     /* Verify VKCS envelope */
     ASSERT_EQ(compressed[0], (char)VKCS_MAGIC_0) << "magic V";
@@ -1323,7 +1323,7 @@ TEST_F(CompressionTest, compressRioFlushMidStream) {
     /* Verify the entire stream decompresses correctly */
     sds compressed = buffer_rio.io.buffer.ptr;
     size_t compressed_len = sdslen(compressed);
-    ASSERT_GT(compressed_len, VKCS_ENVELOPE_SIZE);
+    ASSERT_GT(compressed_len, (size_t)VKCS_ENVELOPE_SIZE);
 
     streamDecompressor sd;
     ASSERT_EQ(streamDecompressorInit(&sd, ALGO_LZ4), 0);
